# backend/api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: https://localhost:8443
    description: Generated server url
...
paths:
...
  /api/v1/events/:
    get:
      tags:
        - Event API
      summary: List events
      description:
        Returns paginated published events with optional search and tag
        filters.
      operationId: getEvents
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: tags
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  type: object
                  additionalProperties:
                    type: object
    post:
      tags:
        - Event API
      summary: Create event
      description:
        "Creates an event from multipart data, optionally including image\
        \ and tags."
      operationId: createEvent
      parameters:
        - name: tagNames
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                event:
                  $ref: "#/components/schemas/EventDTO"
                imageFile:
                  type: string
                  format: binary
              required:
                - event
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
...
  /api/v1/courses/:
    get:
      tags:
        - Course API
      summary: List courses
      description:
        Returns paginated published courses with optional search and tag
        filters.
      operationId: getCourses
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: tags
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  type: object
                  additionalProperties:
                    type: object
    post:
      tags:
        - Course API
      summary: Create course
      description:
        "Creates a course from multipart data, optionally including image\
        \ and tags."
      operationId: createCourse
      parameters:
        - name: tagNames
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                course:
                  $ref: "#/components/schemas/CourseDTO"
                imageFile:
                  type: string
                  format: binary
              required:
                - course
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
...
  /api/v1/admin/users:
    get:
      tags:
        - Admin API
      summary: List users
      description:
        Returns paginated users for admin management with optional text
        filtering.
      operationId: getUsers
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserDTO"
...
  /api/v1/admin/orders:
    get:
      tags:
        - Admin API
      summary: List orders
      description: Returns paginated orders sorted by date for the admin panel.
      operationId: getOrders
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderDTO"
  /api/v1/admin/events:
    get:
      tags:
        - Admin API
      summary: List events
      description:
        Returns paginated events for admin management with optional text
        filtering.
      operationId: getEvents_1
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventDTO"
  /api/v1/admin/courses:
    get:
      tags:
        - Admin API
      summary: List courses
      description:
        Returns paginated courses for admin management with optional text
        filtering.
      operationId: getCourses_1
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CourseDTO"
  /api/v1/admin/reviews/{id}:
    delete:
      tags:
        - Admin API
      summary: Delete review
      description: Deletes a review by id.
      operationId: deleteReview
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
components:
  schemas:
...
    EventDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        priceCents:
          type: integer
          format: int32
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        capacity:
          type: integer
          format: int32
        category:
          type: string
        status:
          type: string
          enum:
            - DRAFT
            - PENDING_REVIEW
            - PUBLISHED
            - ARCHIVED
    CourseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        shortDescription:
          type: string
        priceCents:
          type: integer
          format: int32
        status:
          type: string
          enum:
            - DRAFT
            - PENDING_REVIEW
            - PUBLISHED
            - ARCHIVED
        language:
          type: string
...
    OrderDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        totalAmountCents:
          type: integer
          format: int32
        totalAmountEuros:
          type: string
        status:
          type: string
          enum:
            - PENDING
            - PAID
            - FAILED
            - REFUNDED
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemDTO"
    OrderItemDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        course:
          $ref: "#/components/schemas/CourseDTO"
        event:
          $ref: "#/components/schemas/EventDTO"
        priceAtPurchaseCents:
          type: integer
          format: int32
        priceInEuros:
          type: string
        subscription:
          type: boolean
...
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        isActive:
          type: boolean
        isSubscribed:
          type: boolean
...